Add EIP-2612 permit support; rename to BosonAuthorizedTransferForwarder
The forwarder now supports two off-chain authorization flavours that share
the same action-signature anti-front-run mechanism:

  - ERC-3009 receiveWithAuthorization (existing, used by USDC etc.)
  - EIP-2612 permit (new, used by most DeFi tokens)

Renamed accordingly: BosonERC3009Forwarder -> BosonAuthorizedTransferForwarder
(file, contract, EIP-712 domain). Action typehashes renamed for symmetry:

  DepositFundsAction      -> DepositFundsWithAuthorization
  CommitToOfferAction     -> CommitToOfferWithAuthorization
  (new)                      DepositFundsWithPermit
  (new)                      CommitToOfferWithPermit

Two new entry points: depositFundsWithPermit, commitToOfferWithPermit. They
follow the same per-call shape as the ERC-3009 path: verify action sig,
pull tokens, forceApprove(protocol, value), call protocol, reset allowance.

Permit DoS resilience: the permit() call is wrapped in try/catch. A
front-runner can extract the permit signature from the mempool and consume
the nonce before the forwarder tx lands. When that happens the resulting
allowance is already in place — only the second permit() reverts on stale
nonce, which we ignore. The subsequent safeTransferFrom either succeeds
against the existing allowance or reverts naturally.

Added mock contracts/mock/Foreign20WithPermit.sol (extends OZ ERC20Permit)
for the permit-flow tests. Existing Foreign20WithAuthorization is unchanged.

Tests: 41 cases — happy/sad for both flows, front-run resistance for both,
permit DoS resilience, inner-authorization invariants, and allowance/balance
hygiene.
